Added Color Scheme Toggle option to the website.
Added a Color Scheme Context that'll keep track of the color scheme for the website.
Modified the SCSS files of all the components to adapt to changes in the color scheme.
Added more content to the Devlog Post JSON data file.
Added Color Scheme Toggle button to the Navigation Bar.